Moved canvas into another component
